+++ doc/users/next_whats_new/more-cairo.rst
@@ -1,0 +1,5 @@
+Cairo rendering for Tk canvases
+-------------------------------
+
+The new ``TkCairo`` backend allows Tk canvases to use Cairo rendering instead
+of Agg.
