+++ pytest.ini
@@ -35,7 +35,7 @@ pep8ignore =
     matplotlib/backends/backend_ps.py E202 E203 E225 E228 E231 E261 E262 E271 E301 E302 E303 E401 E402 E501 E701
     matplotlib/backends/backend_svg.py E203 E221 E225 E228 E231 E261 E302 E401 E501
     matplotlib/backends/backend_template.py E302 E303
-    matplotlib/backends/backend_tkagg.py E201 E202 E203 E222 E225 E231 E251 E271 E301 E302 E303 E401 E501 E701 W293
+    matplotlib/backends/_backend_tk.py E201 E202 E203 E222 E225 E231 E251 E271 E301 E302 E303 E401 E501 E701 W293
     matplotlib/backends/tkagg.py E231 E302 E701
     matplotlib/backends/windowing.py E301 E302
     matplotlib/backend_bases.py E225 E712
